requirements.txt
anthropic>=0.8.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
pyyaml>=6.0.0